requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.7
django-crontab==0.7.1
idna==2.10
psycopg2==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.3
selenium~=3.141.0
beautifulsoup4~=4.9.3
djangorestframework~=3.12.2
gunicorn==20.0.4
